♻️ refactor: HomeBanner 내 컴포넌트 분리 #413
- HomePage 내 UI 레이아웃 변경에 따라 인사 문구 등을 출력하는 HomeGreetingSection, 이력서 완성도를 보여주는ResumeHelperBanner, 사용자 유형별 퀵메뉴를 제공하는 HomeMenu를 묶어 HomeTopSection 컴포넌트로 제공
- 기존 광고 배너 캐러셀을 제공했던 HomeBanner를 HomeBannerCarousel로 변경, HomePage에서 단독 호출
